requirements.txt
fastapi==0.115.0
psycopg2==2.9.9
sqlalchemy==2.0.35
uvicorn==0.30.6
python-dotenv==1.0.1